Logs: Integrate in Sentry Client (#2920)

+++ dart/test/sentry_envelope_item_test.dart
@@ -113,7 +113,7 @@ void main() {
           level: SentryLogLevel.info,
           body: 'test2',
           attributes: {
-            'test2': SentryLogAttribute.integer(9001),
+            'test2': SentryLogAttribute.int(9001),
           },
         ),
       ];
